Restricted execution of commands to members with `dev` role
Refactored the commands and events registration, so that we no longer rely on loading files from disk and TypeScript accepts all of the typing.
