+++ .gitignore
@@ -1,6 +1,3 @@
-# Typescript build folder
-built/
-
 # Logs
 logs
 *.log
