+++ events/index.ts
@@ -1,0 +1,7 @@
+import { InteractionCreateEvent } from './interactionCreate';
+import { ReadyEvent } from './ready';
+
+export const AllEvents = [
+  InteractionCreateEvent,
+  ReadyEvent,
+]
