+++ slashCommands/index.ts
@@ -1,0 +1,17 @@
+import { AddAuthorCommand } from './addAuthor'
+import { AddBlockchainCommand } from './addBlockchain'
+import { AddCategoryCommand } from './addCategory'
+import { AddContributorCommand } from './addContributor'
+import { AddGlossaryCommand } from './addGlossary'
+import { AddResourceCommand } from './addResource'
+import { AddTagCommand } from './addTag'
+
+export const AllCommands = [
+  AddAuthorCommand,
+  AddBlockchainCommand,
+  AddCategoryCommand,
+  AddContributorCommand,
+  AddGlossaryCommand,
+  AddResourceCommand,
+  AddTagCommand,
+]
