+++ types/index.ts
@@ -1,1 +1,3 @@
+export * from './events';
 export * from './types';
+export * from './slashCommandConfig'
