+++ types/slashCommandConfig.ts
@@ -1,0 +1,9 @@
+import { RESTPostAPIApplicationCommandsJSONBody } from 'discord-api-types';
+import { CommandInteraction } from 'discord.js';
+
+export interface ISlashCommandConfig {
+  name: string;
+  roles: string[];
+  commandJSON: () => RESTPostAPIApplicationCommandsJSONBody;
+  execute: (interaction: CommandInteraction) => Promise<void>;
+}
